+++ py/desisim/simexp.py
@@ -484,7 +484,8 @@ def simulate_spectra(wave, flux, fibermap=None, obsconditions=None, redshift=Non
     # source types are sky elg lrg qso bgs star , they
     # are only used in specsim.fiberloss for the desi.instrument.fiberloss_method="table" method
 
-    desi.instrument.fiberloss_method = 'fastsim'
+    if specsim_config_file == "desi":
+        desi.instrument.fiberloss_method = 'fastsim'
 
     log.debug('running simulation with {} fiber loss method'.format(desi.instrument.fiberloss_method))
 
@@ -621,7 +622,8 @@ def _specsim_config_for_wave(wave, dwave_out=None, specsim_config_file = "desi")
 
     config.instrument.cameras.b.constants.output_pixel_size = "{:.3f} Angstrom".format(dwave_out)
     config.instrument.cameras.r.constants.output_pixel_size = "{:.3f} Angstrom".format(dwave_out)
-    config.instrument.cameras.z.constants.output_pixel_size = "{:.3f} Angstrom".format(dwave_out)
+    if specsim_config_file == "desi":
+        config.instrument.cameras.z.constants.output_pixel_size = "{:.3f} Angstrom".format(dwave_out)
 
     config.update()
     return config
@@ -825,7 +827,7 @@ def get_mock_spectra(fiberassign, mockdir=None, nside=64, obscon=None):
     return flux, wave, astropy.table.Table(meta), objmeta
 
 def read_mock_spectra(truthfile, targetids, mockdir=None):
-    '''
+    r'''
     Reads mock spectra from a truth file
 
     Args:
@@ -837,6 +839,7 @@ def read_mock_spectra(truthfile, targetids, mockdir=None):
         flux[nspec, nwave]: flux in 1e-17 erg/s/cm2/Angstrom
         wave[nwave]: wavelengths in Angstroms
         truth[nspec]: metadata truth table
+        objtruth: dictionary keyed by objtype type with type-specific truth
     '''
     if len(targetids) != len(np.unique(targetids)):
         from desiutil.log import get_logger
@@ -858,10 +861,19 @@ def read_mock_spectra(truthfile, targetids, mockdir=None):
 
         if 'OBJTYPE' in truth.dtype.names:
             # output of desisim.obs.new_exposure
-            objtype = [oo.decode('ascii').strip().upper() for oo in truth['OBJTYPE']]
+            if isinstance(truth['OBJTYPE'][0], bytes):
+                objtype = [oo.decode('ascii').strip().upper() \
+                        for oo in truth['OBJTYPE']]
+            else:
+                objtype = [oo.strip().upper() for oo in truth['OBJTYPE']]
         else:
             # output of desitarget.mock.build.write_targets_truth
-            objtype = [oo.decode('ascii').strip().upper() for oo in truth['TEMPLATETYPE']]
+            if isinstance(truth['TEMPLATETYPE'][0], bytes):
+                objtype = [oo.decode('ascii').strip().upper() \
+                        for oo in truth['TEMPLATETYPE']]
+            else:
+                objtype = [oo.strip().upper() for oo in truth['TEMPLATETYPE']]
+
         for obj in set(objtype):
             extname = 'TRUTH_{}'.format(obj)
             if extname in fx: